Refactor lookup into a top-level processor
