+++ src/processor/elasticsearch.rs
@@ -1,9 +1,10 @@
 pub mod cluster_settings;
 pub mod index_settings;
 pub mod index_stats;
-pub mod lookup;
 pub mod metadata;
 pub mod nodes;
 pub mod nodes_stats;
 pub mod searchable_snapshots_stats;
 pub mod tasks;
+
+use super::lookup::elasticsearch as lookup;
